requirements.txt
numpy~=1.23.3
streamlit~=1.12.2
streamlit-option-menu==0.3.2
streamlit-lottie==0.0.3	
requests==2.28.1
pandas~=1.4.4
config~=0.5.1	
pyttsx3~=2.90
Pillow~=9.2.0	
